Guardiphy-app/requirements.txt
aiofiles==24.1.0
fastapi==0.115.5
google_auth_oauthlib==1.2.1
httpx==0.27.2
protobuf==5.28.3
pydantic==2.9.2
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.2.0
Requests==2.32.3
starlette==0.41.2
uvicorn==0.32.0
